Creating separate parent class for route presenters
